third_party/WebKit/LayoutTests/media/video-currentTime-before-have-metadata-media-fragment-uri.html

<!DOCTYPE html>
<title>Test currentTime values when setting while readystate is HAVE_NOTHING for media fragment URI.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = "content/test.ogv" + "#t=2";
    assert_equals(video.currentTime, 0);
    video.currentTime = 1;

    video.onloadedmetadata = t.step_func(function() {
        assert_equals(video.currentTime, 1);
    });

    video.onseeked = t.step_func_done();
});
</script>
